+++ app/gen-server/lib/homedb/Interfaces.ts
@@ -3,6 +3,9 @@ import { UserOptions } from "app/common/UserAPI";
 import { Group } from "app/gen-server/entity/Group";
 import { EntityManager } from "typeorm";
 import * as roles from 'app/common/roles';
+import { Organization } from "app/gen-server/entity/Organization";
+import { Workspace } from "app/gen-server/entity/Workspace";
+import { Document } from "app/gen-server/entity/Document";
 
 export interface QueryResult<T> {
   status: number;
@@ -16,5 +19,11 @@ export interface GetUserOptions {
   userOptions?: UserOptions;
 }
 
+export interface UserProfileChange {
+  name?: string;
+  isFirstTimeUser?: boolean;
+}
+
 export type NonGuestGroup = Group & { name: roles.NonGuestRole };
 
+export type Resource = Organization|Workspace|Document;
